Use the clone_url property to clone GitHub repositories
GitHub deprecated use of the unencrypted Git protocol this year per:
https://github.blog/2021-09-01-improving-git-protocol-security-github/#no-more-unauthenticated-git
This results in a quiet failure when cloning GitHub repositories. The
clone_url property of a GitHub Repository object contains the HTTPS URL
that can be used to clone a repository and will likely continue to be
the source for a valid URL to clone a GitHub repository successfully.