Constrict support to Python 3.6 and newer
This is the oldest version of Python currently tested. Though this
version has already reached End of Life it still sees use on older
operating system releases so I did not want to remove support as part
of this effort. The version of `setuptools` that is required is the
first version with support for the `python_requires` kwarg for the
setuptools.setup() function.