Add the ability to configure GitHub session timeouts
Add the ability to modify the timeouts available for GitHub sessions in
the github3.py package. When running the scraper against a GitHub org
with a large number of repositories a read timeout was observed which
would crash the run. With this implementation and modifying the read
timeout the scraper was able to process the same GitHub org with no
issues.

Please see https://requests.readthedocs.io/en/latest/user/advanced/#timeouts
for more information about these timeouts.